+++ examples/pushstate-server.ts
@@ -28,9 +28,7 @@ app.route({
 					: {},
 			);
 		}
-		await reply.code(result.statusCode)
-			.headers(result.headers)
-			.send(result.stream);
+		await result.send(reply.raw);
 	},
 });