+++ test/http.spec.ts
@@ -47,17 +47,23 @@ function shouldHaveHeader(header: string) {
 
 function multipartHandler(res: request.Response, cb: (err: Error | null, body: unknown) => void) {
 	const chunks: Buffer[] = [];
-	res.on('data', chunk => {
-		chunks.push(<Buffer> chunk);
+	let length = 0;
+	res.on('data', (chunk: Buffer | string) => {
+		const buffer = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk);
+		chunks.push(buffer);
+		length += buffer.length;
 	});
 	let end = false;
-	res.on('error', err => {
+	res.on('error', (err: Error) => {
 		end = true;
 		cb(err, null);
 	});
 	res.on('end', () => {
 		end = true;
-		res.text = Buffer.concat(chunks).toString();
+		const concatChunks = Buffer.concat(chunks, length);
+		res.text = Buffer.isEncoding(res.charset)
+			? concatChunks.toString(res.charset)
+			: concatChunks.toString();
 		cb(null, res.text);
 	});
 	res.on('close', () => {